Swap `--format=plain` and `--format=text`
